crm_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
scikit-learn==1.3.2
textblob==0.17.1
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
